requirements.txt
langchain
langchain-community
PyMuPDF==1.22.3
pytesseract==0.3.9
Pillow==11.2.1
ocrmypdf==16.7.0
pandas==2.3.0
torch>=2.0.0
transformers>=4.3.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.3
easyocr>=1.7
opencv-python-headless>=4.7.0
llama-cpp-python
dearpygui>=1.10.1
